+++ tests/integrations/stdlib/test_httplib.py
@@ -240,3 +240,103 @@ def test_outgoing_trace_headers_head_sdk(sentry_init, monkeypatch):
         assert sorted(request_headers["baggage"].split(",")) == sorted(
             expected_outgoing_baggage_items
         )
+
+
+@pytest.mark.parametrize(
+    "trace_propagation_targets,host,path,trace_propagated",
+    [
+        [
+            [],
+            "example.com",
+            "/",
+            True,
+        ],
+        [
+            ["https://example.com/"],
+            "example.com",
+            "/",
+            True,
+        ],
+        [
+            ["https://example.com/"],
+            "example.com",
+            "",
+            False,
+        ],
+        [
+            ["https://example.com"],
+            "example.com",
+            "",
+            True,
+        ],
+        [
+            ["https://example.com", r"https?:\/\/[\w\-]+(\.[\w\-]+)+\.net"],
+            "example.net",
+            "",
+            False,
+        ],
+        [
+            ["https://example.com", r"https?:\/\/[\w\-]+(\.[\w\-]+)+\.net"],
+            "good.example.net",
+            "",
+            True,
+        ],
+        [
+            ["https://example.com", r"https?:\/\/[\w\-]+(\.[\w\-]+)+\.net"],
+            "good.example.net",
+            "/some/thing",
+            False,
+        ],
+        [
+            ["https://example.com", r"https?:\/\/[\w\-]+(\.[\w\-]+)+\.net\/.*"],
+            "good.example.net",
+            "/some/thing",
+            True,
+        ],
+    ],
+)
+def test_option_trace_propagation_targets(
+    sentry_init, monkeypatch, trace_propagation_targets, host, path, trace_propagated
+):
+    # HTTPSConnection.send is passed a string containing (among other things)
+    # the headers on the request. Mock it so we can check the headers, and also
+    # so it doesn't try to actually talk to the internet.
+    mock_send = mock.Mock()
+    monkeypatch.setattr(HTTPSConnection, "send", mock_send)
+
+    sentry_init(
+        trace_propagation_targets=trace_propagation_targets,
+        traces_sample_rate=1.0,
+    )
+
+    headers = {
+        "baggage": (
+            "sentry-trace_id=771a43a4192642f0b136d5159a501700, "
+            "sentry-public_key=49d0f7386ad645858ae85020e393bef3, sentry-sample_rate=0.01337, "
+        )
+    }
+
+    transaction = Transaction.continue_from_headers(headers)
+
+    with start_transaction(
+        transaction=transaction,
+        name="/interactions/other-dogs/new-dog",
+        op="greeting.sniff",
+        trace_id="12312012123120121231201212312012",
+    ) as transaction:
+
+        HTTPSConnection(host).request("GET", path)
+
+        (request_str,) = mock_send.call_args[0]
+        request_headers = {}
+        for line in request_str.decode("utf-8").split("\r\n")[1:]:
+            if line:
+                key, val = line.split(": ")
+                request_headers[key] = val
+
+        if trace_propagated:
+            assert "sentry-trace" in request_headers
+            assert "baggage" in request_headers
+        else:
+            assert "sentry-trace" not in request_headers
+            assert "baggage" not in request_headers
