+++ tests/tracing/test_misc.py
@@ -6,6 +6,7 @@ import os
 import sentry_sdk
 from sentry_sdk import Hub, start_span, start_transaction, set_measurement
 from sentry_sdk.tracing import Span, Transaction
+from sentry_sdk.tracing_utils import should_propagate_trace
 
 try:
     from unittest import mock  # python 3.3 and above
@@ -271,3 +272,49 @@ def test_set_meaurement_public_api(sentry_init, capture_events):
     (event,) = events
     assert event["measurements"]["metric.foo"] == {"value": 123, "unit": ""}
     assert event["measurements"]["metric.bar"] == {"value": 456, "unit": "second"}
+
+
+@pytest.mark.parametrize(
+    "url, expected_propagation_decision",
+    [
+        ("http://example.com", True),
+        ("http://example.com/", False),
+        ("https://example.com", False),
+        ("http://www.example.com", False),
+        ("https://example.com:8080/foo?bar=baz", False),
+        ("http://bla.net", False),
+        ("https://bla.net", False),
+        ("http://www.bla.net", True),
+        ("https://www.bla.net", True),
+    ],
+)
+def test_should_propagate_trace(url, expected_propagation_decision):
+    trace_propagation_targets = [
+        "http://example.com",
+        r"https?:\/\/[\w\-]+(\.[\w\-]+)+\.net",  # matches http://bla.blub.net but not http://blub.net
+    ]
+
+    assert (
+        should_propagate_trace(url, trace_propagation_targets)
+        == expected_propagation_decision
+    )
+
+
+@pytest.mark.parametrize(
+    "url, expected_propagation_decision",
+    [
+        ("http://example.com", True),
+        ("https://example.com", True),
+        ("https://example.com:8080/foo?bar=baz", True),
+        ("http://bla.net", True),
+        ("http://www.bla.net", True),
+        ("http://evil.hacker.org/", True),
+    ],
+)
+def test_should_propagate_trace_empty(url, expected_propagation_decision):
+    trace_propagation_targets = []
+
+    assert (
+        should_propagate_trace(url, trace_propagation_targets)
+        == expected_propagation_decision
+    )
